Updated plugin data, updated control panel.

+++ Empty/Data/IPluginData.cs
@@ -5,6 +5,6 @@ namespace SDRSharp.FreqToProscan
     public interface IPluginData
     {
         List<string> ProscanLines { get; }
-        List<float> Frequencies { get; }
+        List<int> Frequencies { get; }
     }
 }
